requirements.txt
streamlit>=1.24.0
openai>=1.0.0
supabase>=1.0.3
python-dotenv>=1.0.0
requests>=2.31.0
psutil>=5.9.0
asyncio>=3.4.3
crawl4ai>=1.0.0
httpx>=0.24.1
beautifulsoup4>=4.12.2
markdown>=3.4.3
python-markdown-math>=0.8 